Move lowering into codegen package
Lowering is not important enough to deserve it's own
dedicated build step. This commit does not change how
lowering and codegen are run but we might integrate
them tighter together in the future.
