+++ tools/src/test/scala/scala/scalanative/FrameworkTest.scala
@@ -4,7 +4,7 @@ import nir.Global
 
 import org.scalatest._
 
-class FrameworkTest extends CodeGenSpec with Matchers {
+class FrameworkTest extends codegen.CodeGenSpec with Matchers {
 
   "The test framework" should "return the definitions for a single class" in {
     link("A$",
